+++ data/src/main/java/com/sopetit/data/service/AuthService.kt
@@ -3,7 +3,7 @@ package com.sopetit.data.service
 import com.sopetit.data.base.BaseResponse
 import com.sopetit.data.base.EndPoints
 import com.sopetit.data.entity.request.LogInRequestDto
-import com.sopetit.data.entity.response.LogInResponseDto
+import com.sopetit.data.entity.response.auth.LogInResponseDto
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
